会話の吹き出しの位置関係を変える
