+++ lib/ui/screens/ai_screen/voice_screen.dart
@@ -8,22 +8,6 @@ import 'package:aitrip/ui/components/voice_container.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class SpeechState {
-  final String lastWords;
-  final bool isListening;
-  final bool isSpeechEnabled;
-
-  SpeechState({
-    this.lastWords = '',
-    this.isListening = false,
-    this.isSpeechEnabled = false,
-  });
-}
-
-final speechProvider = StateNotifierProvider<SpeechNotifier, SpeechState>(
-  (ref) => SpeechNotifier(),
-);
-
 class VoiceScreen extends ConsumerStatefulWidget {
   const VoiceScreen({super.key});
 
@@ -58,7 +42,7 @@ class VoiceScreenState extends ConsumerState<VoiceScreen> {
     }
 
     final isLoading = ref.watch(isLoadingProvider);
-    final hotelOption = ref.watch(hotelOptionProvider);
+    ref.watch(hotelOptionProvider);
     return Scaffold(
       body: Center(
         child: Column(
@@ -66,6 +50,7 @@ class VoiceScreenState extends ConsumerState<VoiceScreen> {
             Padding(
               padding: const EdgeInsets.all(30.0),
               child: Container(
+                // グレーの背景
                 height: MediaQuery.of(context).size.height * 0.3,
                 width: MediaQuery.of(context).size.width * 0.9,
                 decoration: BoxDecoration(
@@ -78,40 +63,42 @@ class VoiceScreenState extends ConsumerState<VoiceScreen> {
                 child: Padding(
                   padding: const EdgeInsets.only(
                       left: 30.0, right: 30.0, bottom: 30.0),
-                  child: Column(
-                    children: [
-                      AnimatedSwitcher(
-                        duration: const Duration(milliseconds: 500),
-                        switchInCurve: Curves.easeIn,
-                        switchOutCurve: Curves.easeOut,
-                        transitionBuilder:
-                            (Widget child, Animation<double> animation) {
-                          // フェードインとスライドインのアニメーションを組み合わせる
-                          return FadeTransition(
-                            opacity: animation,
-                            child: SlideTransition(
-                              position: Tween<Offset>(
-                                begin: const Offset(0, 0.5),
-                                end: Offset.zero,
-                              ).animate(animation),
-                              child: child,
-                            ),
-                          );
-                        },
-                        //声を聞き取っている場合、ServerContainerを表示、そうでない場合はUserContainerを表示
-                        child: isLoading
-                            ? const LoadingContainer()
-                            : speechState.isListening
-                                ? UserContainer(
-                                    key: const ValueKey('user'),
-                                    text: speechState.lastWords,
-                                  )
-                                : ServerContainer(
-                                    key: const ValueKey('server'),
-                                    text: message != null ? message.text : '',
-                                  ),
-                      ),
-                    ],
+                  child: SingleChildScrollView(
+                    child: Column(
+                      children: [
+                        AnimatedSwitcher(
+                          duration: const Duration(milliseconds: 500),
+                          switchInCurve: Curves.easeIn,
+                          switchOutCurve: Curves.easeOut,
+                          transitionBuilder:
+                              (Widget child, Animation<double> animation) {
+                            // フェードインとスライドインのアニメーションを組み合わせる
+                            return FadeTransition(
+                              opacity: animation,
+                              child: SlideTransition(
+                                position: Tween<Offset>(
+                                  begin: const Offset(0, 0.5),
+                                  end: Offset.zero,
+                                ).animate(animation),
+                                child: child,
+                              ),
+                            );
+                          },
+                          //声を聞き取っている場合、ServerContainerを表示、そうでない場合はUserContainerを表示
+                          child: isLoading
+                              ? const LoadingContainer()
+                              : speechState.isListening
+                                  ? UserContainer(
+                                      key: const ValueKey('user'),
+                                      text: speechState.lastWords,
+                                    )
+                                  : ServerContainer(
+                                      key: const ValueKey('server'),
+                                      text: message != null ? message.text : '',
+                                    ),
+                        ),
+                      ],
+                    ),
                   ),
                 ),
               ),
